Rename searchRowTest to searchRow
